+++ src/main/java/de/ostfalia/view/StoreForm.java
@@ -49,7 +49,7 @@ public class StoreForm implements Serializable {
         Form form = new Form();
         String id = FacesContext.getCurrentInstance().getExternalContext().getRequestParameterMap().get("id");
         if(id != null){
-            this.store = ss.findById(Long.valueOf(id));
+            this.store = ss.findById(Integer.valueOf(id));
         } else {
             this.store = new Store();
         }
